Recenter Y Axis + Workaround for apparent Wait Time bug
Recenter Y Axis + Workaround for apparent Wait Time bug.

The Wait Time and Recentering Time appear not to be fully honored by the CinemachineFreeLook camera, in particular, it seems that we need to increase the Wait Time by about 5, so we used 6 since we want a wait time of about 1 second.