Added horizonal recentering with R (keyboard) and L3 (gamepad)
Added horizonal recentering with R (keyboard) and L3 (gamepad).

There is a drawback: if the Follow Camera needs recentering both horizontally and vertically, and the player presses the R key / L3 button, it will correctly trigger a smooth and gradual horizontal recentering, but the vertical recentering will appear to happen all at once.