Added left and right aim cameras (no animations yet)
Added left and right aim cameras (no animations yet.)

With the F key or the R3 button on a Gamepad the player can swap the default aim camera.

With the Right mouse button or the Left Trigger on a Gamepad, the player can switch from the Follow Camera to the default Aim Camera.

NOTE: that the Y Axis Recentering timing has been changed for all 3 virtual cameras.